requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
python-dotenv==1.1.0
pydantic==2.11.2
python-jose[cryptography]==3.3.0 